+++ contracts/fund/src/testing/migration/migration_scenario.rs
@@ -1,10 +1,8 @@
 use crate::contract::{CONTRACT_NAME, CONTRACT_VERSION};
 
-use cosmwasm_std::{coin, Decimal};
-use interface::fund::{
-    ConfigResponse, ExtensionQueryMsg, InstantiateMsg, MigrateMsg, QueryMsg,
-    VaultenatorExtensionQueryMsg,
-};
+use cosmwasm_schema::cw_serde;
+use cosmwasm_std::{coin, Decimal, Uint128};
+use interface::fund::{ExtensionQueryMsg, MigrateMsg, QueryMsg, VaultenatorExtensionQueryMsg};
 use neutron_test_tube::{Account, Module, Runner, Wasm};
 use osmosis_std::types::cosmwasm::wasm::v1::{
     MsgMigrateContract, MsgMigrateContractResponse, QueryContractInfoRequest,
@@ -16,13 +14,43 @@ use testing::{
     utils::store_code,
 };
 
+#[cw_serde]
+pub struct V010InstantiateMsg {
+    pub admin: String,         // manages contract configuration
+    pub controller: String,    // manages contract balance sheet
+    pub treasury: String,      // account fees are paid to
+    pub strategy_cap: Uint128, // maximum value of strategy deposits
+    pub float: Decimal,        // percentage of balance sheet that can be withdrawn
+    pub token0: String,
+    pub token1: Option<String>,
+    pub management_fee_rate: Decimal,
+    pub performance_fee_rate: Decimal,
+    pub vault_type: String,
+}
+
+#[cw_serde]
+pub struct V010ConfigResponse {
+    pub admin: String,
+    pub controller: String,
+    pub treasury: String,
+    pub strategy_cap: Uint128,
+    pub float: Option<Decimal>,
+    pub strategy_denom: String,
+    pub token0: String,
+    pub token1: Option<String>,
+    pub management_fee_rate: Decimal,
+    pub performance_fee_rate: Decimal,
+    pub estimate_cycle_profit: Option<Decimal>,
+    pub vault_type: String,
+}
+
 #[test]
 fn test_migration() {
     let env = TestEnv::new();
     let wasm = Wasm::new(&env.app);
 
     let wasm_byte_code =
-        std::fs::read("../../contracts/fund/src/testing/artifacts/fund_vault-v010.wasm").unwrap();
+        std::fs::read("../../contracts/fund/src/testing/migration/fund-v010.wasm").unwrap();
 
     let fund_vault_v003 = wasm
         .store_code(&wasm_byte_code, None, &env.signer)
@@ -30,7 +58,7 @@ fn test_migration() {
         .data
         .code_id;
 
-    let msg = InstantiateMsg {
+    let msg = V010InstantiateMsg {
         admin: env.signer.address().to_string(),
         controller: env.signer.address().to_string(),
         treasury: env.treasury.address().to_string(),
@@ -65,9 +93,8 @@ fn test_migration() {
         VaultenatorExtensionQueryMsg::Config {},
     ));
 
-    let config: ConfigResponse = wasm.query(&contract_addr, &query_msg).unwrap();
+    let config: V010ConfigResponse = wasm.query(&contract_addr, &query_msg).unwrap();
 
-    // note: for v0.0.2 onwards the denom is omitted
     let expected_strategy_denom = format!("factory/{}/fund", contract_addr);
 
     assert_eq!(config.strategy_denom, expected_strategy_denom);
@@ -88,7 +115,10 @@ fn test_migration() {
                 sender: env.signer.address(),
                 contract: contract_addr.clone(),
                 code_id,
-                msg: serde_json_wasm::to_vec(&MigrateMsg {}).unwrap(),
+                msg: serde_json_wasm::to_vec(&MigrateMsg {
+                    redemption_contract: env.controller.address().to_string(),
+                })
+                .unwrap(),
             },
             "/cosmwasm.wasm.v1.MsgMigrateContract",
             &env.signer,
@@ -118,6 +148,10 @@ fn test_migration() {
 
     assert_eq!(config.strategy_denom, expected_strategy_denom);
     assert_eq!(config.token0, BASE_DENOM.to_string());
+    assert_eq!(
+        config.redemption_contract,
+        env.controller.address().to_string()
+    );
 
     let version = env.query_version_fund(&wasm, &contract_addr).unwrap();