+++ contracts/redemption/src/errors.rs
@@ -1,0 +1,32 @@
+use cosmwasm_std::StdError;
+use thiserror::Error;
+
+#[derive(Error, Debug, PartialEq)]
+pub enum ContractError {
+    #[error("{0}")]
+    Std(#[from] StdError),
+
+    #[error("Unable to perform migration")]
+    MigrationError {},
+
+    #[error("No redemptions found")]
+    NoRedemptionsFound {},
+
+    #[error("Unauthorized")]
+    Unauthorized {},
+
+    #[error("Vault not set")]
+    VaultNotSet {},
+
+    #[error("Unauthorized fund")]
+    UnauthorizedFund {},
+
+    #[error("Insufficient funds")]
+    InsufficientFunds {},
+}
+
+impl ContractError {
+    pub fn generic_err(msg: impl Into<String>) -> ContractError {
+        ContractError::Std(StdError::generic_err(msg))
+    }
+}